requirements.txt
python-telegram-bot==19.7
yt-dlp
ffmpeg-python